Docstring cleanup
